requirements.txt
fastapi
uvicorn
discord.py
apscheduler
python-dotenv
jinja2
aiohttp
python-multipart
